Use CSRF protection
The nosurf library is being used, so that we can create and check
for a CSRF token in order to prevent CSRF attacks.

Therefore I added a hidden field to all forms and added the nosurf
middleware to all routes, because we have the signout form potentially
on all sites.